fix proposal details stop rendering
